deleting files

+++ controllers/admin.js
@@ -1,5 +1,8 @@
 const mongoose = require("mongoose");
 const Product = require("../models/product");
+
+const fileHelper = require('../util/file')
+
 const { validationResult } = require("express-validator/check");
 
 exports.getAddProduct = (req, res, next) => {
@@ -138,6 +141,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -173,15 +177,25 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
+  Product.findById(prodId).then(product => {
+    console.log(`product`, product)
 
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => {
-      const error = new Error("Creating a product failed");
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+    if(!product) {
+      next(new Error('Product not found'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
+  .then(() => {
+    console.log("DESTROYED PRODUCT");
+    
+    res.redirect("/admin/products");
+  })
+  .catch((err) => {
+    const error = new Error("Creating a product failed");
+    error.httpStatusCode = 500;
+    return next(error);
+  });
+  
+    
 };
